Remove uncessary code for FrameRate and Improved the PlayerMovement Script and the start MenuInGame
I just simplify the code that it necessary, I just said you know what just limiting to 60fps is more than enough, I don't need to put a fancy slider that controls the frame rate in the menu Options. and  added a limit of the speed and so started working on the Menu that will be shown when you are alreading playing the game.